try to fix shader vertex lighting issue. fix issue where ftenq protocol positions were sticking. fixed manifest file windowless infinite wait. Added support for ^[\img\foo^] images on the console if they are at the start of the line. Console objects may now be drawn by csqc. Key input is blocked for now, lest csqc get access to things its not allowed to access. Will add provisions later. fixed sound menu crash. added ultrawide speex codec. clamp micamp levels to try to boost audio quality/audiability. rtp control plugins are now allowed to use opus, speex support is correctly queried instead of assumed. WriteFloat now works properly, and not just for ents. no longer sending prespawn/precaches/signon commands to clients that do not expect them. fix sql column header queries. more verbose about active nq protocol verions. -condebug overrides config.
git-svn-id: https://svn.code.sf.net/p/fteqw/code/trunk@4461 fc73d0e0-1445-4013-8a0c-d673dee63da5

+++ engine/client/vid.h
@@ -45,7 +45,7 @@ extern rendererstate_t currentrendererstate;
 
 typedef struct vrect_s
 {
-	int				x,y,width,height;
+	float				x,y,width,height;
 } vrect_t;
 
 typedef struct
